created domain app domain and migrate domain related logic to  that app. Updated imports where applicable
